backend/requirements.txt
boto==2.38.0
click==5.1
Django==1.8.4
django-cacheops==2.4.1
django-cors-headers==1.1.0
django-crispy-forms==1.5.1
django-extensions==1.5.5
django-filter==0.11.0
django-hstore==1.4
django-imagekit==3.2.7
django-localflavor-us==1.1
django-object-actions==0.5.1
django-push-notifications==1.3.1
django-storages-redux==1.3
django-suit==0.2.14
djangorestframework==3.1.3
djangorestframework-gis==0.9.3
djangorestframework-hstore==1.3
djoser==0.3.1
drf-nested-routers==0.10.0        
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
jsonfield==1.0.3
livereload==2.4.0
Markdown==2.6.2
MarkupSafe==0.23
mccabe==0.3.1
mkdocs==0.14.0
pdfkit==0.5.0
pexpect==4.0.1
pickleshare==0.5
pilkit==1.1.13
Pillow==2.9.0
psycopg2==2.6
ptyprocess==0.5
pyflakes==0.9.2
PySocks==1.5.6
python-dateutil==2.4.2
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.7.0
rest-condition==1.0.1
sendgrid==1.4.0
sendgrid-django==1.4.0
simplegeneric==0.8.1
six==1.9.0
smtpapi==0.2.0
sqlparse==0.1.16
tornado==4.2.1
traitlets==4.0.0
twilio==4.5.0
wheel==0.24.0
